Write a Java program to count the number of days between two given years.﻿
Input format 
The input consists of two integers .start year and end year. 
Output format 
The output prints the number of days between two years and if the end year is greater than the start year then it has to print "End year must be greater than the first year !".